requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
distro==1.9.0
exceptiongroup==1.2.2
face-recognition==1.3.0
Flask==2.2.0
flask-cors==5.0.1
Flask-Session==0.8.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==4.12.0
itsdangerous==2.1.1
Jinja2==3.1.0
jiter==0.8.2
MarkupSafe==2.1.1
mkl-service==2.4.0
msgspec==0.19.0
numpy==1.26.4
openai==1.64.0
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
oracledb==2.5.1
Pillow==8.4.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==2.2.0
zipp==3.9.0




